% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_bayes_glm_spec}
\alias{define_bayes_glm_spec}
\title{Define Bayesian GLM Model Specification}
\usage{
define_bayes_glm_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a Bayesian GLM model specification using the `stan` engine for regression tasks.
}
