% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_c5_0_spec}
\alias{define_c5_0_spec}
\title{Define C5.0 Model Specification}
\usage{
define_c5_0_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"classification"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a C5.0 decision tree model specification using the `C5.0` engine for classification tasks.
}
