% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_decision_tree_spec.R
\name{define_decision_tree_spec}
\alias{define_decision_tree_spec}
\title{Define Decision Tree Model Specification}
\usage{
define_decision_tree_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Decision Tree Model Specification
}
