% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_elastic_net_spec}
\alias{define_elastic_net_spec}
\title{Define Elastic Net Model Specification}
\usage{
define_elastic_net_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates an elastic net model specification using the `glmnet` engine for regression tasks.
}
