% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_gbm_spec}
\alias{define_gbm_spec}
\title{Define Gradient Boosting Machine (GBM) Model Specification}
\usage{
define_gbm_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a GBM model specification using the `xgboost` engine for classification or regression tasks.
}
