% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_glmboost_spec}
\alias{define_glmboost_spec}
\title{Define GLMBoost Model Specification}
\usage{
define_glmboost_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a GLMBoost model specification using the `glmboost` engine for regression tasks.
}
