% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_knn_spec}
\alias{define_knn_spec}
\title{Define k-Nearest Neighbors Model Specification}
\usage{
define_knn_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a k-NN model specification using the `kknn` engine for classification or regression tasks.
}
