% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_logistic_regression_spec.R
\name{define_logistic_regression_spec}
\alias{define_logistic_regression_spec}
\title{Define Logistic Regression Model Specification}
\usage{
define_logistic_regression_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type ("classification").}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Logistic Regression Model Specification
}
