% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_naive_bayes_spec.R
\name{define_naive_bayes_spec}
\alias{define_naive_bayes_spec}
\title{Define Naive Bayes Model Specification}
\usage{
define_naive_bayes_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type ("classification").}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Naive Bayes Model Specification
}
