% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_neural_network_spec.R
\name{define_neural_network_spec}
\alias{define_neural_network_spec}
\title{Define Neural Network Model Specification (nnet)}
\usage{
define_neural_network_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Neural Network Model Specification (nnet)
}
