% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_penalized_logistic_regression_spec}
\alias{define_penalized_logistic_regression_spec}
\title{Define Penalized Logistic Regression Model Specification}
\usage{
define_penalized_logistic_regression_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"classification"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a penalized logistic regression model specification using the `glmnet` engine for classification tasks.
}
