% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_pls_spec}
\alias{define_pls_spec}
\title{Define Partial Least Squares Model Specification}
\usage{
define_pls_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a PLS model specification using the `mixOmics` engine for regression tasks.
}
