% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_qda_spec}
\alias{define_qda_spec}
\title{Define Quadratic Discriminant Analysis Model Specification}
\usage{
define_qda_spec(task)
}
\arguments{
\item{task}{Character string specifying the task type ("classification").}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Quadratic Discriminant Analysis Model Specification
}
