% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_random_forest_spec}
\alias{define_random_forest_spec}
\title{Define Random Forest Model Specification}
\usage{
define_random_forest_spec(task, train_data)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}

\item{train_data}{A data frame containing the training data (used to determine the number of predictors).}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a random forest model specification using the `ranger` engine for classification or regression tasks.
}
