% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_ridge_regression_spec}
\alias{define_ridge_regression_spec}
\title{Define Ridge Regression Model Specification}
\usage{
define_ridge_regression_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a ridge regression model specification using the `glmnet` engine for regression tasks.
}
