% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_svm_radial_spec}
\alias{define_svm_radial_spec}
\title{Define SVM Radial Model Specification}
\usage{
define_svm_radial_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define SVM Radial Model Specification
}
