% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.fastml_model.R
\name{explain}
\alias{explain}
\title{Explain model (Generic)}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{An object to explain.}

\item{...}{Additional arguments passed to the method.}
}
\description{
A generic function to provide explanations for models.
Different objects can have specialized methods, such as \code{explain.fastml_model}.
}
