\name{framingham}
\docType{data}
\alias{framingham}
\title{Framingham Heart Study Data}
\description{
  This dataset is derived from the Framingham Heart Study and contains various clinical and demographic variables used to predict coronary heart disease risk over a ten-year period.
}
\format{\describe{
  \item{male}{Integer indicator for male sex.}
  \item{age}{Participant age in years.}
  \item{education}{Education level.}
  \item{currentSmoker}{Whether the participant currently smokes.}
  \item{cigsPerDay}{Number of cigarettes smoked per day.}
  \item{BPMeds}{Whether blood pressure medication is used.}
  \item{prevalentStroke}{History of stroke at baseline.}
  \item{prevalentHyp}{History of hypertension at baseline.}
  \item{diabetes}{Diabetes diagnosis.}
  \item{totChol}{Total cholesterol.}
  \item{sysBP}{Systolic blood pressure.}
  \item{diaBP}{Diastolic blood pressure.}
  \item{BMI}{Body mass index.}
  \item{heartRate}{Heart rate.}
  \item{glucose}{Glucose level.}
  \item{TenYearCHD}{Ten year risk of coronary heart disease.}
}}
