% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_helpers.R
\name{get_best_workflows}
\alias{get_best_workflows}
\title{Get Best Workflows}
\usage{
get_best_workflows(models, best_model_name)
}
\arguments{
\item{models}{A nested list of model workflows.}

\item{best_model_name}{A named character vector of chosen best engines.}
}
\description{
Extracts the best workflows from a nested list of model workflows based on the provided best model names.
}
