% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fastml_model.R
\name{plot.fastml_model}
\alias{plot.fastml_model}
\title{Plot Function for fastml_model}
\usage{
\method{plot}{fastml_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fastml_model}.}

\item{...}{Additional arguments (not used).}
}
\value{
Displays comparison plots of model performances.
}
\description{
Generates plots to compare the performance of different models.
}
