% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_stability.R
\name{plot.fastml_stability}
\alias{plot.fastml_stability}
\title{Plot method for fastml_stability objects}
\usage{
\method{plot}{fastml_stability}(x, top_n = 15, ...)
}
\arguments{
\item{x}{A fastml_stability object from explain_stability().}

\item{top_n}{Integer. Number of top features to display. Default is 15.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for fastml_stability objects
}
