% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fastml_model.R
\name{predict.fastml_model}
\alias{predict.fastml_model}
\title{Predict Function for fastml_model}
\usage{
\method{predict}{fastml_model}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{fastml_model}.}

\item{newdata}{A data frame containing new data for prediction.}

\item{...}{Additional arguments (not used).}
}
\value{
A vector of predictions.
}
\description{
Makes predictions on new data using the trained model.
}
