% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_lime.R
\name{predict_model.model_fit}
\alias{predict_model.model_fit}
\title{Internal predict_model method for parsnip fits}
\usage{
predict_model.model_fit(x, newdata, type, ...)
}
\description{
Shim for parsnip model objects so that lime's predict_model generic ignores
unused arguments passed via `...`.
}
\keyword{internal}
