% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.fastml.R
\name{save.fastml}
\alias{save.fastml}
\title{Save Model Function}
\usage{
save.fastml(model, filepath)
}
\arguments{
\item{model}{An object of class \code{fastml}.}

\item{filepath}{A string specifying the file path to save the model.}
}
\value{
No return value, called for its side effect of saving the model object to a file.
}
\description{
Saves the trained model object to a file.
}
