% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_model}
\alias{save_model}
\title{Save Model Function}
\usage{
save_model(model, filepath)
}
\arguments{
\item{model}{An object of class \code{fastml_model}.}

\item{filepath}{A string specifying the file path to save the model.}
}
\description{
Saves the trained model object to a file.
}
