% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tuneGrid.R
\name{validate_tuneGrid}
\alias{validate_tuneGrid}
\title{Validate and Complete the tuneGrid}
\usage{
validate_tuneGrid(tuneGrid, default_params, required_params, is_cv_enabled)
}
\arguments{
\item{tuneGrid}{User-provided tuning grid.}

\item{default_params}{Default hyperparameter ranges.}

\item{required_params}{Required hyperparameters for the algorithm.}

\item{is_cv_enabled}{Logical indicating whether cross-validation is enabled.}
}
\value{
A validated and possibly modified tuneGrid.
}
\description{
Ensures that the tuneGrid includes all required hyperparameters and adjusts it based on cross-validation.
}
