% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_bagging_spec.R
\name{define_bagging_spec}
\alias{define_bagging_spec}
\title{Define Bagging Model Specification}
\usage{
define_bagging_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Bagging Model Specification
}
