% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_bayes_glm_spec.R
\name{define_bayes_glm_spec}
\alias{define_bayes_glm_spec}
\title{Define Bayesian GLM Model Specification}
\usage{
define_bayes_glm_spec(task)
}
\arguments{
\item{task}{Character string specifying the task type ("regression").}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Bayesian GLM Model Specification
}
