% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_c5_0_spec.R
\name{define_c5_0_spec}
\alias{define_c5_0_spec}
\title{Define C5.0 Model Specification}
\usage{
define_c5_0_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define C5.0 Model Specification
}
