% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_decision_tree_spec}
\alias{define_decision_tree_spec}
\title{Define Decision Tree Model Specification}
\usage{
define_decision_tree_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a decision tree model specification using the `rpart` engine for classification or regression tasks.
}
