% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_deep_learning_spec}
\alias{define_deep_learning_spec}
\title{Define Deep Learning Model Specification with Keras}
\usage{
define_deep_learning_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a deep learning model specification using the `keras` engine for classification or regression tasks.
}
