% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_elastic_net_spec.R
\name{define_elastic_net_spec}
\alias{define_elastic_net_spec}
\title{Define Elastic Net Model Specification}
\usage{
define_elastic_net_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type ("regression").}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Elastic Net Model Specification
}
