% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_knn_spec}
\alias{define_knn_spec}
\title{Define K-Nearest Neighbors Model Specification}
\usage{
define_knn_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type: "classification" or "regression".}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define K-Nearest Neighbors Model Specification
}
