% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_lda_spec.R
\name{define_lda_spec}
\alias{define_lda_spec}
\title{Define Linear Discriminant Analysis Model Specification}
\usage{
define_lda_spec(task)
}
\arguments{
\item{task}{Character string specifying the task type ("classification").}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Linear Discriminant Analysis Model Specification
}
