% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_lightgbm_spec}
\alias{define_lightgbm_spec}
\title{Define LightGBM Model Specification}
\usage{
define_lightgbm_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a LightGBM model specification using the `lightgbm` engine for classification or regression tasks.
}
