% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_linear_regression_spec}
\alias{define_linear_regression_spec}
\title{Define Linear Regression Model Specification}
\usage{
define_linear_regression_spec(task)
}
\arguments{
\item{task}{Character string specifying the task type ("regression").}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Linear Regression Model Specification
}
