% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_logistic_regression_spec}
\alias{define_logistic_regression_spec}
\title{Define Logistic Regression Model Specification}
\usage{
define_logistic_regression_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"classification"`.}
}
\value{
A list containing the model specification (`model_spec`), default parameters (`default_params`), and required parameters (`required_params`).
}
\description{
Creates a logistic regression model specification using the `glm` engine for classification tasks.
}
