% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_naive_bayes_spec}
\alias{define_naive_bayes_spec}
\title{Define Naive Bayes Model Specification}
\usage{
define_naive_bayes_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"classification"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a Naive Bayes model specification using the `klaR` engine for classification tasks.
}
