% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_neural_network_spec}
\alias{define_neural_network_spec}
\title{Define Neural Network Model Specification}
\usage{
define_neural_network_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a neural network model specification using the `nnet` engine for classification or regression tasks.
}
