% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_pls_spec.R
\name{define_pls_spec}
\alias{define_pls_spec}
\title{Define Partial Least Squares Model Specification}
\usage{
define_pls_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type ("regression").}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Partial Least Squares Model Specification
}
