% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_qda_spec}
\alias{define_qda_spec}
\title{Define Quadratic Discriminant Analysis Model Specification}
\usage{
define_qda_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task. Must be `"classification"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates a QDA model specification using the `MASS` engine for classification tasks.
}
