% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_ridge_regression_spec}
\alias{define_ridge_regression_spec}
\title{Define Ridge Regression Model Specification}
\usage{
define_ridge_regression_spec(task, tune = FALSE)
}
\arguments{
\item{task}{Character string specifying the task type ("regression").}

\item{tune}{Logical indicating whether to use tuning parameters.}
}
\value{
List containing the model specification (`model_spec`).
}
\description{
Define Ridge Regression Model Specification
}
