% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{define_svm_radial_spec}
\alias{define_svm_radial_spec}
\title{Define SVM with Radial Basis Function Kernel Model Specification}
\usage{
define_svm_radial_spec(task)
}
\arguments{
\item{task}{A character string specifying the type of task: `"classification"` or `"regression"`.}
}
\value{
A list containing the model specification, default parameters, and required parameters.
}
\description{
Creates an SVM model specification with an RBF kernel using the `kernlab` engine for classification or regression tasks.
}
