% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainer_helpers.R
\name{fastml_prepare_explainer_inputs}
\alias{fastml_prepare_explainer_inputs}
\title{Internal helper to prepare explainer inputs from a fastml object}
\usage{
fastml_prepare_explainer_inputs(object)
}
\description{
Internal helper to prepare explainer inputs from a fastml object
}
\keyword{internal}
