% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_helpers.R
\name{flatten_and_rename_models}
\alias{flatten_and_rename_models}
\title{Flatten and Rename Models}
\usage{
flatten_and_rename_models(models)
}
\arguments{
\item{models}{A named list of model objects, optionally nested by engine.}
}
\description{
Flatten and Rename Models
}
