% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_helpers.R
\name{get_best_model_idx}
\alias{get_best_model_idx}
\title{Get Best Model Indices by Metric and Group}
\usage{
get_best_model_idx(df, metric, group_cols = c("Model", "Engine"))
}
\description{
Get Best Model Indices by Metric and Group
}
