% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.fastml.R
\name{load_model}
\alias{load_model}
\title{Load Model Function}
\usage{
load_model(filepath)
}
\arguments{
\item{filepath}{A string specifying the file path to load the model from.}
}
\value{
An object of class \code{fastml}.
}
\description{
Loads a trained model object from a file.
}
