% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fastml_model.R
\name{predict.fastml_model}
\alias{predict.fastml_model}
\title{Predict Function for fastml_model}
\usage{
\method{predict}{fastml_model}(object, newdata, type = "auto", ...)
}
\arguments{
\item{object}{An object of class \code{fastml_model}.}

\item{newdata}{A data frame containing new data for prediction.}

\item{type}{Type of prediction. Default is \code{"auto"}, which returns class labels for classification and numeric predictions for regression.
Other options include \code{"prob"} for class probabilities (classification only).}

\item{...}{Additional arguments (not used).}
}
\value{
A vector or data frame of predictions.
}
\description{
Makes predictions on new data using the trained model.
}
