% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fastml_model.R
\name{summary.fastml_model}
\alias{summary.fastml_model}
\title{Summary Function for fastml_model}
\usage{
\method{summary}{fastml_model}(object, sort_metric = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{fastml_model}.}

\item{sort_metric}{A string specifying which metric to sort the models by.
Default is \code{NULL}, which prioritizes the optimized metric.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a concise and user-friendly summary of the models' performances.
}
\description{
Provides a concise and user-friendly summary of the model's performances.
}
